+++ src/main/webui/src/app/main/doctor/patients/patients.routes.ts
@@ -43,7 +43,7 @@ export const PATIENTS_ROUTES: Routes = [
 			},
 			{
 				path: '**',
-				loadComponent: () => import('src/app/shared/not-found').then(m => m.NotFoundComponent)
+				loadComponent: () => import('@shared/not-found').then(m => m.NotFoundComponent)
 			}
 		]
 	}