+++ src/test/java/eu/valawai/c0_patient_treatment_ui/persistence/TreatmentValueFeedbackEntityTest.java
@@ -8,9 +8,9 @@
 
 package eu.valawai.c0_patient_treatment_ui.persistence;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
@@ -121,14 +121,14 @@ public class TreatmentValueFeedbackEntityTest {
 					final TreatmentValueFeedbackPayload payload = (TreatmentValueFeedbackPayload) asserter
 							.getData("PAYLOAD");
 					final long now = (Long) asserter.getData("NOW");
-					assertTrue("Unexpected reported feedback time", now <= feedback.createdTime);
+					assertTrue(now <= feedback.createdTime, "Unexpected reported feedback time");
 					assertEquals(payload.value_name, feedback.valueName);
 					assertEquals(payload.alignment, feedback.alignment, 0.0000001d);
 					break;
 				}
 			}
 
-			assertTrue("Not found stored action in the treatment", found);
+			assertTrue(found, "Not found stored action in the treatment");
 		});
 	}