+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # C0_patient_treatment_ui
 
+## Version 1.2.0 (February 20, 2026)
 
+ - Fix development environment to use the new docker image,
+  because openjdk is deprecated.
+ - Round alignment value.
+
+ 
 ## Version 1.1.0 (April 30, 2025)
 
  - Use another method for pulling data.